+++ b2sdk/v1/bucket.py
@@ -8,13 +8,30 @@
 #
 ######################################################################
 
+import functools
+
 from typing import Optional
+from .file_version import file_version_info_from_new_file_version, FileVersionInfoFactory, FileVersionInfo
 from b2sdk import _v2 as v2
 from b2sdk.utils import validate_b2_file_name
 
 
-# Overridden to retain the obsolete copy_file and start_large_file methods
+def translate_single_file_version(func):
+    @functools.wraps(func)
+    def inner(*a, **kw):
+        file_version = file_version_info_from_new_file_version(func(*a, **kw))
+        if not isinstance(file_version, FileVersionInfo):
+            return file_version_info_from_new_file_version(file_version)
+        return file_version
+
+    return inner
+
+
+# Overridden to retain the obsolete copy_file and start_large_file methods and return old FileVersionInfo from
+# several methods
 class Bucket(v2.Bucket):
+    FILE_VERSION_FACTORY = staticmethod(FileVersionInfoFactory)
+
     def copy_file(
         self,
         file_id,
@@ -64,6 +81,9 @@ class Bucket(v2.Bucket):
         return self.api.services.large_file.start_large_file(
             self.id_, file_name, content_type=content_type, file_info=file_info
         )
+
+    _create_file = translate_single_file_version(v2.Bucket._create_file)
+    copy = translate_single_file_version(v2.Bucket.copy)
 
 
 class BucketFactory(v2.BucketFactory):